Change session management to TOCTOU bug
Previously, an in use sessionId could be pruned if the user requested their
history before answering a question. This would cause unrecoverable errors on
the client side.

To prevent this, we now generate a session id on request, but do not persist it
to the database until an answer comes in. Only then do we unshift the session
into our user schema.